Moved openSearchItem, findInPage functionality to our `CoreReaderFragment` as we were using same code twice for these methods in app and custom module.
